# Writerside/api/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/admin/assignChoice:
    put:
      tags:
      - admin-assign-choice-controller
      description: Assigning ChoiceClass to student
      operationId: assignChoice
      parameters:
      - name: choiceClassId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: ChoiceClass could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Assigned ChoiceClass to student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentChoicesResponse'
    post:
      tags:
      - admin-assign-choice-controller
      description: Assigning alternate Choice to student
      operationId: assignAlternateChoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlternateChoiceRequest'
        required: true
      responses:
        "200":
          description: Assigned alternate Choice to student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentChoicesResponse'
        "404":
          description: Student could not be found.
          content:
            application/json:
              schema:
                type: string
    delete:
      tags:
      - admin-assign-choice-controller
      description: Deleting ChoiceClass selection from student
      operationId: deleteChoiceSelection
      parameters:
      - name: choiceClassId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Deleting ChoiceClass selection from student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentChoicesResponse'
        "404":
          description: ChoiceClass could not be found.
          content:
            application/json:
              schema:
                type: string
  /api/admin/admin:
    get:
      tags:
      - admin-admin-controller
      description: Gets Admin
      operationId: getAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Found Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
    put:
      tags:
      - admin-admin-controller
      description: Edit Admin
      operationId: editAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
      - admin-admin-controller
      description: Signs up new Admin
      operationId: createAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "201":
          description: Signed up new Admin
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-admin-controller
      description: Deletes Admin
      operationId: deleteAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Deleted Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
...
  /api/admin/lesson:
    post:
      tags:
      - admin-lesson-controller
      description: Creating new lesson
      operationId: createLesson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonCreationRequest'
        required: true
      responses:
        "400":
          description: Lesson is not available.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Created new lesson
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TapeResponse'
        "404":
          description: Tape could not be found.
          content:
            application/json:
              schema:
                type: string
    delete:
      tags:
      - admin-lesson-controller
      description: Deleting lesson
      operationId: deleteLesson
      parameters:
      - name: lessonId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Lesson could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Deleted lesson
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TapeResponse'
...
  /api/admin/studentChoices:
    get:
      tags:
      - admin-assign-choice-controller
      description: Getting choices of student
      operationId: getStudentChoices
      parameters:
      - name: studentId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found Choices of Student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentChoicesResponse'
        "404":
          description: Student could not be found.
          content:
            application/json:
              schema:
                type: string
...
  /api/admin/classesStudents:
    get:
      tags:
      - admin-assign-choice-controller
      description: Get all classes with students
      operationId: getClassesWithStudents
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Found all classes with Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassStudentsResponse'
  /api/admin/choiceTapes:
    get:
      tags:
      - admin-assign-choice-controller
      description: Get all tapes by year
      operationId: getTapes
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Found all tapes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChoiceTapeResponse'
...
  /api/admin/alternativeChoice:
    delete:
      tags:
      - admin-assign-choice-controller
      description: Deleting alternate Choice from student
      operationId: deleteAlternativeChoiceClass
      parameters:
      - name: choiceClassId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: ChoiceClass could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Deleted alternate Choice from student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentChoicesResponse'
components:
  schemas:
...
    ChoiceResponse:
      type: object
      properties:
        choiceNumber:
          type: integer
          format: int32
        classChoiceResponses:
          type: array
          items:
            $ref: '#/components/schemas/ClassChoiceResponse'
    ClassChoiceResponse:
      type: object
      properties:
        classId:
          type: integer
          format: int64
        choiceClassId:
          type: integer
          format: int64
        selected:
          type: boolean
        name:
          type: string
        tapeId:
          type: integer
          format: int64
        teacherResponse:
          $ref: '#/components/schemas/TeacherResponse'
    RuleResponse:
      type: object
      properties:
        name:
          type: string
        subjectResponses:
          type: array
          items:
            $ref: '#/components/schemas/SubjectResponse'
    SubjectResponse:
      type: object
      properties:
        name:
          type: string
...
    StudentChoicesResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        studentId:
          type: integer
          format: int64
        studentClassId:
          type: integer
          format: int64
        name:
          type: string
        year:
          type: integer
          format: int32
        ruleResponses:
          type: array
          items:
            $ref: '#/components/schemas/RuleResponse'
        choiceResponses:
          type: array
          items:
            $ref: '#/components/schemas/ChoiceResponse'
    TeacherResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        teacherId:
          type: integer
          format: int64
        abbreviation:
          type: string
    AdminSignupRequest:
      type: object
      properties:
        firstname:
          type: string
        surname:
          type: string
...
    LessonCreationRequest:
      type: object
      properties:
        day:
          type: integer
          format: int32
        hour:
          type: integer
          format: int32
        tapeId:
          type: integer
          format: int64
    LessonResponse:
      type: object
      properties:
        lessonId:
          type: integer
          format: int64
        day:
          type: integer
          format: int32
        hour:
          type: integer
          format: int32
    TapeResponse:
      type: object
      properties:
        tapeId:
          type: integer
          format: int64
        name:
          type: string
        year:
          type: integer
          format: int32
        releaseYear:
          type: integer
          format: int32
        lessonResponses:
          type: array
          items:
            $ref: '#/components/schemas/LessonResponse'
        lk:
          type: boolean
...
    AlternateChoiceRequest:
      type: object
      properties:
        classId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
...
    ClassStudentsResponse:
      type: object
      properties:
        name:
          type: string
        teacherResponse:
          $ref: '#/components/schemas/TeacherResponse'
        studentSurveillanceResponses:
          type: array
          items:
            $ref: '#/components/schemas/StudentSurveillanceResponse'
        tapeName:
          type: string
        sizeWarning:
          type: boolean
        sizeCritical:
          type: boolean
    StudentSurveillanceResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        studentId:
          type: integer
          format: int64
        studentClassId:
          type: integer
          format: int64
        name:
          type: string
        year:
          type: integer
          format: int32
    ChoiceTapeClassResponse:
      type: object
      properties:
        classId:
          type: integer
          format: int64
        name:
          type: string
        teacherResponse:
          $ref: '#/components/schemas/TeacherResponse'
    ChoiceTapeResponse:
      type: object
      properties:
        tapeId:
          type: integer
          format: int64
        name:
          type: string
        choiceTapeClassResponses:
          type: array
          items:
            $ref: '#/components/schemas/ChoiceTapeClassResponse'
        choiceTapeClassResponse:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/ChoiceTapeClassResponse'
...
    AdminResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        adminId:
          type: integer
          format: int64
...